requirements.txt
python_dateutil==2.9.0.post0
Requests==2.31.0
TeslaPy==2.9.0
